+++ oidc_provider/tests/cases/test_provider_info_endpoint.py
@@ -1,14 +1,14 @@
+from unittest.mock import patch
+
 from django.core.cache import cache
-from mock import patch
+from django.test import RequestFactory
+from django.test import TestCase
+from django.test import override_settings
 
 try:
     from django.urls import reverse
 except ImportError:
     from django.core.urlresolvers import reverse
-
-from django.test import RequestFactory
-from django.test import TestCase
-from django.test import override_settings
 
 from oidc_provider.views import ProviderInfoView